feat(image-spec): Expose StaticFile and OsFile
